requirements.txt
fastapi==0.115.12
psycopg2_binary==2.9.10
supabase==2.15.2
